hourglass-realm/backend/requirements.txt
fastapi
langchain
langchain-google-genai
pillow
python-dotenv
uvicorn